// packages/desktop-app/sdk/packages/core/src/synai-core/types.ts
export interface SynAICoreStartInput
	extends Omit<StartSessionInput, "config" | "localRuntime"> {
	config: SynAICoreStartConfig;
	localRuntime?: LocalRuntimeStartOptions;
}